* v2-master: (29 commits)
  Fix console image name
  Fix not detecting admin user is admin scope is not stratos.admin
  Update status_updates.md
  Update passthrough.go
  Update roadmap
  Weekly status updates
  Remove auto select of service and improve service select
  Add delete space confirmation @ space level
  Removed console.log, fixed console.log linting, show warning for space pills as well
  Ignore cookie domain for now
  Fix file path for output
  Trigger on tag
  Update release pipeline
  Fix bad error formatting
  Tidy up pagination key creation - Ensure there's one common base way to create pagination key - Tidy up how that base is used WRT to entity relation pagination key - Ensure all lists that run against all endpoints have a common pagination key
  Fix space selection when user returns to CFOrgSpace selector
  Fix issue where circular reference failed to denormalize an app's space - Connect to scf, no other endpoint (bug is app specific) - to reproduce...   - Fresh load on CF page   - Click Services nav item   - Click Applications nav item   - apps in third column did not denormalize their space entity, causing undefined error - Caused by app's space's service instances containing app bindings
  Unit test fixes
  Fix unit tests
  Minor scss fix for code climate
  ...

+++ src/frontend/app/shared/components/list/list-types/cf-spaces/cf-space-card/cf-space-card.component.ts
@@ -2,27 +2,24 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable, Subscription } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
-
 import { ISpace } from '../../../../../../core/cf-api.types';
+import { CurrentUserPermissions } from '../../../../../../core/current-user-permissions.config';
 import { CurrentUserPermissionsService } from '../../../../../../core/current-user-permissions.service';
-import { EntityServiceFactory } from '../../../../../../core/entity-service-factory.service';
 import { getSpaceRolesString } from '../../../../../../features/cloud-foundry/cf.helpers';
-import {
-  CloudFoundryEndpointService,
-} from '../../../../../../features/cloud-foundry/services/cloud-foundry-endpoint.service';
-import {
-  CloudFoundryOrganizationService,
-} from '../../../../../../features/cloud-foundry/services/cloud-foundry-organization.service';
+import { CloudFoundryEndpointService } from '../../../../../../features/cloud-foundry/services/cloud-foundry-endpoint.service';
+import { CloudFoundryOrganizationService } from '../../../../../../features/cloud-foundry/services/cloud-foundry-organization.service';
 import { RouterNav } from '../../../../../../store/actions/router.actions';
 import { AppState } from '../../../../../../store/app-state';
+import { entityFactory, spaceSchemaKey } from '../../../../../../store/helpers/entity-factory';
 import { APIResource } from '../../../../../../store/types/api.types';
 import { EndpointUser } from '../../../../../../store/types/endpoint.types';
 import { CfUserService } from '../../../../../data-services/cf-user.service';
+import { ComponentEntityMonitorConfig } from '../../../../../shared.types';
+import { ConfirmationDialogConfig } from '../../../../confirmation-dialog.config';
+import { ConfirmationDialogService } from '../../../../confirmation-dialog.service';
 import { MetaCardMenuItem } from '../../../list-cards/meta-card/meta-card-base/meta-card.component';
 import { CardCell } from '../../../list.types';
-import { CurrentUserPermissions } from '../../../../../../core/current-user-permissions.config';
-import { ComponentEntityMonitorConfig } from '../../../../../shared.types';
-import { entityFactory, spaceSchemaKey } from '../../../../../../store/helpers/entity-factory';
+
 
 @Component({
   selector: 'app-cf-space-card',
@@ -53,7 +50,8 @@ export class CfSpaceCardComponent extends CardCell<APIResource<ISpace>> implemen
     private cfEndpointService: CloudFoundryEndpointService,
     private store: Store<AppState>,
     private cfOrgService: CloudFoundryOrganizationService,
-    private currentUserPermissionsService: CurrentUserPermissionsService
+    private currentUserPermissionsService: CurrentUserPermissionsService,
+    private confirmDialog: ConfirmationDialogService
   ) {
     super();
   }
@@ -157,11 +155,19 @@ export class CfSpaceCardComponent extends CardCell<APIResource<ISpace>> implemen
   }
 
   delete = () => {
-    this.cfOrgService.deleteSpace(
-      this.spaceGuid,
-      this.orgGuid,
-      this.cfEndpointService.cfGuid
+    const confirmation = new ConfirmationDialogConfig(
+      'Delete Space',
+      `Are you sure you want to delete space '${this.row.entity.name}'?`,
+      'Delete',
+      true
     );
+    this.confirmDialog.open(confirmation, () => {
+      this.cfOrgService.deleteSpace(
+        this.spaceGuid,
+        this.orgGuid,
+        this.cfEndpointService.cfGuid
+      );
+    });
   }
 
   goToSummary = () => this.store.dispatch(new RouterNav({
